refactor: Move class files
