+++ feature/explore/src/main/java/com/teamoffroad/feature/explore/presentation/ExploreRoute.kt
@@ -1,4 +1,4 @@
-package com.teamoffroad.feature.explore
+package com.teamoffroad.feature.explore.presentation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.runtime.Composable
