+++ feature/explore/src/main/java/com/teamoffroad/feature/explore/presentation/ExploreViewModel.kt
@@ -1,0 +1,15 @@
+package com.teamoffroad.feature.explore.presentation
+
+import android.location.Location
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import javax.inject.Inject
+
+@HiltViewModel
+class ExploreViewModel @Inject constructor() : ViewModel() {
+
+    private val _location = MutableStateFlow<Location?>(null)
+    val location: StateFlow<Location?> = _location
+}
