+++ feature/explore/src/main/java/com/teamoffroad/feature/explore/presentation/navigation/ExploreNavigation.kt
@@ -1,4 +1,4 @@
-package com.teamoffroad.feature.explore.navigation
+package com.teamoffroad.feature.explore.presentation.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
@@ -6,7 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.teamoffroad.core.navigation.MainTabRoute
-import com.teamoffroad.feature.explore.ExploreRoute
+import com.teamoffroad.feature.explore.presentation.ExploreRoute
 
 fun NavController.navigateExplore(navOptions: NavOptions) {
     navigate(MainTabRoute.Explore, navOptions)
